+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Changed
+
+- Flash messages in User Settings and Admin Panel are now dismissed on click
+
 ## [0.5] - 2021-05-24
 
 ### Added
@@ -129,6 +135,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - User cache for reduced DB queries
 - Page title updates to reflect current view
 
+[Unreleased]: https://github.com/APB9785/phxBB/compare/0.5...HEAD
 [0.5]: https://github.com/APB9785/phxBB/compare/0.4.1...0.5
 [0.4.1]: https://github.com/APB9785/phxBB/compare/0.4...0.4.1
 [0.4]: https://github.com/APB9785/phxBB/compare/0.3.1...0.4